v2: return zero values on error
In Go, the default assumption is that if a func returns (T, error) and
its error is non-nil, the T value is invalid unless otherwise
documented. (io.Reader is the common example of a documented case
where both results can be valid)

On the flip side, the best practice for funcs returning (T, error) is
to only set one or the other non-nil, lest callers accidentally depend
on T always being non-nil and ossifying the implementation, making it
unable to be refactored later.

Updates #cleanup

Signed-off-by: Brad Fitzpatrick <bradfitz@tailscale.com>
Signed-off-by: hosom <hosom@github.com>